created format.js and moved trim, trimWeeks, newLine, newLineWeeks with corresponding tests being moved into format_test. all tests passing
