+++ lib/render.js
@@ -1,9 +1,13 @@
 'use strict';
 
+const {
+  trim,
+  newLine,
+} = require('./format')
+
 const { 
   getMonthString,
-  getMonthLength,
-  checkForLeapYear
+  getMonthLength
 } = require('./dateConversion')
 
 module.exports.renderCal = (dayOfWeek, dateObj) => {
@@ -12,12 +16,11 @@ module.exports.renderCal = (dayOfWeek, dateObj) => {
   let fixedHeader = module.exports.createFixedHeader()
   let weeks = module.exports.createWeeks(dayOfWeek, dateObj.month, dateObj.year)
   let cal = []
-  cal.push(module.exports.newLine(module.exports.trim(mainHeader)))
-  cal.push(module.exports.newLine(module.exports.trim(fixedHeader)))
+  cal.push(newLine(trim(mainHeader)))
+  cal.push(newLine(trim(fixedHeader)))
   weeks.forEach(week => {
-    cal.push(module.exports.newLine(module.exports.trim(week)))
+    cal.push(newLine(trim(week)))
   })
-  // cal.push(module.exports.newLineWeeks(module.exports.trimWeeks(weeks[0])))
   return cal
 }
 
@@ -32,7 +35,6 @@ module.exports.renderYearCal = (dayOfWeek, dateObj) => {
   weeks.forEach(week => {
     cal.push(week)
   })
-  // cal.push(module.exports.newLineWeeks(module.exports.trimWeeks(weeks[0])))
   return cal
 }
 
@@ -41,7 +43,6 @@ module.exports.renderCalString = (array) => {
   array.forEach((segment) => {
     string += segment[0]
   })
-  // console.log('renderCalString &&&&&&&&&&&&&&&&&&&&&', string)
   return string
 }
 
@@ -112,30 +113,4 @@ module.exports.createWeeks = (dayOfWeek, month, year) => {
 
   return weeks
 
-}
-
-module.exports.trimWeeks = (array) => {
-  let trimWeek = array.map(week => module.exports.trim(week))
-  return trimWeek
-}
-
-module.exports.trim = (array) => {
-  let trim = []
-  let string = array[0]
-  string = string.trimRight()
-  trim.push(string)
-  return trim
-}
-
-module.exports.newLineWeeks = (array) => {
-  let newLineWeek = array.map(week => module.exports.newLine(week))
-  return newLineWeek
-}
-
-module.exports.newLine = (array) => {
-  let newLine = []
-  let string = array[0]
-  string += '\n'
-  newLine.push(string)
-  return newLine
 }
