+++ lib/year.js
@@ -2,10 +2,12 @@
 
 const {
   centerString,
-  trim,
-  newLine,
   renderYearCal
 } = require('../lib/render')
+const {
+  trim,
+  newLine
+} = require('../lib/format')
 const {
   getDayOfWeek
 } = require('../lib/zeller')
@@ -13,7 +15,8 @@ const {
 module.exports.createYearHeader = (yearString) => {
   let yearHeader = centerString(yearString, 64, true)
   yearHeader = trim(yearHeader)
-  return yearHeader = newLine(newLine(yearHeader))
+  yearHeader = newLine(newLine(yearHeader))
+  return yearHeader
 }
 
 module.exports.createYearArray = (yearString) => {
